pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "git-translate-commits"
version = "1.0.0"
description = "Translate git commit messages to a target language using local or LLM-based translation"
readme = "README.md"
license = {text = "GPL-3.0-or-later"}
requires-python = ">=3.10"
authors = [
    { name = "Fernando Paladini", email = "fernando@paladini.dev" },
]
maintainers = [
    { name = "Fernando Paladini", email = "fernando@paladini.dev" },
]
keywords = ["git", "commit", "translation", "llm", "i18n"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Version Control :: Git",
    "Topic :: Text Processing :: Linguistic",
    "Typing :: Typed",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "gitpython>=3.1.0",
    "argostranslate>=1.9.0",
    "langdetect>=1.0.9",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
llm = [
    "litellm>=1.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.0.0",
    "litellm>=1.0.0",
]

[project.scripts]
git-translate-commits = "git_translate_commits.cli:app"

[project.urls]
Homepage = "https://github.com/paladini/git-translate-commits"
Repository = "https://github.com/paladini/git-translate-commits"
Issues = "https://github.com/paladini/git-translate-commits/issues"
Changelog = "https://github.com/paladini/git-translate-commits/blob/main/CHANGELOG.md"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
git_translate_commits = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
